requirements.txt
sanic>=20.12.0
gino>=1.0.1
sqlalchemy>=1.3.0
marshmallow>=3.0.0
pyjwt>=2.0.0
alembic>=1.0.0
